Enable Gradle Wrapper dependency updates via Renovate
We used a custom workflow to update the wrapper before. When switching to Renovate we kept the workflow because Renovate didn't update the distributionSha256Sum property. It's been able to do that for a while now and the gradle-update action is not maintained anymore, so it's a good oppertunity to switch.

We need to remove the gradlew-update.yaml workflow from the repositories using Gradle.